services/model-server/requirements.txt
# Hunyuan3D-2.1 Model Server
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
torch>=2.1.0
pillow>=10.0.0
trimesh>=4.5.0
numpy>=1.26.0
scipy>=1.12.0
# Note: shap-e is installed via git in the Dockerfile (not on PyPI)
